requirements.txt
# Core dependencies
anthropic>=0.40.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Market data
yfinance>=0.2.36

# HTTP requests
requests>=2.31.0

# LangGraph and LangChain ecosystem
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-anthropic>=0.3.0

# Optional: Development and testing
pytest>=8.0.0
black>=24.0.0
